[UI] Phase 4: consolidate icons and shapes under ui/assets/icons (#18744)
Merges ui/assets/new-icons into ui/assets/icons. Relocates ui/components/shapes
to ui/assets/icons/shapes. Removes the redundant ui/components/icons barrel
left over from Phase 2 #18736 in favor of the single ui/assets/icons home
established by Phase 2 #18730.

Fixes #18744
Refs #18659, #18654
