+++ ui/assets/icons/shapes/RoundedTriangle.tsx
