+++ ui/components/ConfirmationModal.tsx
@@ -23,8 +23,8 @@ import { useLazyPingKubernetesQuery } from '@/rtk-query/connection';
 import { getK8sConfigIdsFromK8sConfig } from '../utils/multi-ctx';
 import React, { useEffect, useState, FC } from 'react';
 import { iconMedium, iconSmall } from '../css/icons.styles';
-import { RoundedTriangleShape } from './shapes/RoundedTriangle';
-import RedOctagonSvg from './shapes/Octagon';
+import { RoundedTriangleShape } from '../assets/icons/shapes/RoundedTriangle';
+import RedOctagonSvg from '../assets/icons/shapes/Octagon';
 import PatternIcon from '../assets/icons/Pattern';
 import { useNotification } from '../utils/hooks/useNotification';
 import { EVENT_TYPES } from '../lib/event-types';
