--- ui/components/icons/index.ts
@@ -1,58 +1,0 @@
-/**
- * Centralized icon barrel.
- *
- * All MUI icon imports should go through this file.
- * When Sistent provides its own icon set, only this file needs to change.
- */
-export {
-  AccessTime,
-  AddCircle,
-  AddCircleOutlined,
-  ArrowBack,
-  ArrowDropDown,
-  BarChart,
-  BuildRounded,
-  Cached,
-  CheckCircle,
-  Close,
-  Code,
-  Delete,
-  DirectionsCar,
-  DoneAll,
-  Edit,
-  Error,
-  ExpandLess,
-  ExpandMore,
-  Explore,
-  FileCopy,
-  FileUpload,
-  Filter,
-  Fullscreen,
-  FullscreenExit,
-  GetApp,
-  GroupAdd,
-  HelpOutline,
-  HelpOutlineOutlined,
-  HelpOutlined,
-  Info,
-  InfoOutlined,
-  Launch,
-  ListAlt,
-  Lock,
-  MergeOutlined,
-  MoreHoriz,
-  MoreVert,
-  OpenInNewOutlined,
-  Public,
-  Publish,
-  Reply,
-  Save,
-  SaveAs,
-  SaveOutlined,
-  Search,
-  Settings,
-  SimCard,
-  SupervisedUserCircle,
-  TouchApp,
-  Warning,
-} from '@mui/icons-material';
